Added Logs stuff

+++ PokeD.SCON.UILibrary/PlayersDataGridModel.cs
@@ -4,7 +4,7 @@
     {
         public int Number { get; set; }
         public string Name { get; set; }
-        public uint GameJoltID { get; set; }
+        public ulong GameJoltID { get; set; }
         public string IP { get; set; }
         public uint Ping { get; set; }
         public bool Online { get; set; }
